refactor:code splitted into reducers,filter utilities and API services
